Revert "Scroll reveal added from react bits"
This reverts commit cd715abc5554933684709041d8f907170e9f417e.
