+++ aavesh/package.json
@@ -13,8 +13,7 @@
     "@splinetool/react-spline": "^4.0.0",
     "emailjs-com": "^3.2.0",
     "framer-motion": "^11.3.23-alpha.3",
-    "gsap": "^3.13.0",
-    "prop-types": "^15.8.1",
+  "gsap": "^3.12.5",
     "react": "^18.3.1",
     "react-dom": "^18.3.1",
     "react-feather": "^2.0.10",
