+++ aavesh/src/components/Footer.jsx
@@ -4,7 +4,6 @@ import emailjs from "emailjs-com";
 import socials from "../Socials";
 import { motion } from "framer-motion";
 import Heading from "./Heading";
-import ScrollReveal from "./ScrollReveal";
 
 const Footer = (props) => {
   const [user, setUser] = useState({
@@ -93,11 +92,7 @@ const Footer = (props) => {
             </div>
             </div>
             <div className="border-t-2 border-cyan-600 mb-4 lg:block hidden"></div>
-            <Heading as="h2" sizeClass="text-5xl md:text-7xl" uppercase={false} colorClass="text-white font-iceland" className="font-bold m-2">
-              <ScrollReveal>
-                LET&apos;S CONNECT!
-              </ScrollReveal>
-            </Heading>
+            <Heading as="h2" sizeClass="text-5xl md:text-7xl" uppercase={false} colorClass="text-white font-iceland" className="font-bold m-2">LET&apos;S CONNECT!</Heading>
             <form
               onSubmit={handleSubmit}
               className="flex flex-col items-center space-y-4"
