--- aavesh/src/components/ScrollReveal.jsx
@@ -1,148 +1,0 @@
-import { useEffect, useRef, useMemo } from 'react';
-import PropTypes from 'prop-types';
-import { gsap } from 'gsap';
-import { ScrollTrigger } from 'gsap/ScrollTrigger';
-
-gsap.registerPlugin(ScrollTrigger);
-
-const ScrollReveal = ({
-  children,
-  scrollContainerRef,
-  enableBlur = true,
-  baseOpacity = 0,
-  baseRotation = 3,
-  blurStrength = 4,
-  containerClassName = "",
-  textClassName = "",
-  rotationEnd = "bottom bottom",
-  wordAnimationEnd = "bottom bottom"
-}) => {
-  const containerRef = useRef(null);
-
-  const splitText = useMemo(() => {
-    if (typeof children !== 'string') {
-      return children;
-    }
-    const text = children;
-    return text.split(/(\s+)/).map((word, index) => {
-      if (word.match(/^\s+$/)) return word;
-      return (
-        <span className="inline-block word" key={index}>
-          {word}
-        </span>
-      );
-    });
-  }, [children]);
-
-  useEffect(() => {
-    const el = containerRef.current;
-    if (!el) return;
-
-    const scroller =
-      scrollContainerRef && scrollContainerRef.current
-        ? scrollContainerRef.current
-        : window;
-
-    gsap.fromTo(
-      el,
-      { transformOrigin: '0% 50%', rotate: baseRotation },
-      {
-        ease: 'none',
-        rotate: 0,
-        scrollTrigger: {
-          trigger: el,
-          scroller,
-          start: 'top bottom',
-          end: rotationEnd,
-          scrub: true,
-        },
-      }
-    );
-
-    const wordElements = el.querySelectorAll('.word');
-
-    if (wordElements.length > 0) {
-      gsap.fromTo(
-        wordElements,
-        { opacity: baseOpacity, willChange: 'opacity' },
-        {
-          ease: 'none',
-          opacity: 1,
-          stagger: 0.05,
-          scrollTrigger: {
-            trigger: el,
-            scroller,
-            start: 'top bottom-=20%',
-            end: wordAnimationEnd,
-            scrub: true,
-          },
-        }
-      );
-
-      if (enableBlur) {
-        gsap.fromTo(
-          wordElements,
-          { filter: `blur(${blurStrength}px)` },
-          {
-            ease: 'none',
-            filter: 'blur(0px)',
-            stagger: 0.05,
-            scrollTrigger: {
-              trigger: el,
-              scroller,
-              start: 'top bottom-=20%',
-              end: wordAnimationEnd,
-              scrub: true,
-            },
-          }
-        );
-      }
-    } else {
-      gsap.fromTo(
-        el,
-        { 
-          opacity: baseOpacity, 
-          filter: enableBlur ? `blur(${blurStrength}px)` : 'blur(0px)',
-          willChange: 'opacity, filter'
-        },
-        {
-          ease: 'none',
-          opacity: 1,
-          filter: 'blur(0px)',
-          scrollTrigger: {
-            trigger: el,
-            scroller,
-            start: 'top bottom-=20%',
-            end: wordAnimationEnd,
-            scrub: true,
-          },
-        }
-      );
-    }
-
-    return () => {
-      ScrollTrigger.getAll().forEach(trigger => trigger.kill());
-    };
-  }, [scrollContainerRef, enableBlur, baseRotation, baseOpacity, rotationEnd, wordAnimationEnd, blurStrength, children]);
-
-  return (
-    <div ref={containerRef} className={containerClassName}>
-      <div className={textClassName}>{splitText}</div>
-    </div>
-  );
-};
-
-ScrollReveal.propTypes = {
-    children: PropTypes.node.isRequired,
-    scrollContainerRef: PropTypes.object,
-    enableBlur: PropTypes.bool,
-    baseOpacity: PropTypes.number,
-    baseRotation: PropTypes.number,
-    blurStrength: PropTypes.number,
-    containerClassName: PropTypes.string,
-    textClassName: PropTypes.string,
-    rotationEnd: PropTypes.string,
-    wordAnimationEnd: PropTypes.string,
-};
-
-export default ScrollReveal;
