+++ aavesh/src/pages/TeamPage.jsx
@@ -3,7 +3,6 @@ import TeamInfoCard from "../components/TeamInfoCard";
 import { CoreTeam, Team } from "../Members";
 import Navbar from "../components/Header";
 import Heading from "../components/Heading";
-import ScrollReveal from "../components/ScrollReveal";
 
 // Temporary hardcoded Faculty data (can move to Members.js)
 const facultyCoordinator = {
@@ -24,11 +23,7 @@ const TeamPage = () => {
         <div className="min-h-screen bg-black text-white flex flex-col p-6">
           
           {/* 🧑‍🏫 Faculty Coordinator Section */}
-          <Heading as="h1" size="lg" className="mb-8 ml-28">
-            <ScrollReveal>
-              FACULTY COORDINATOR
-            </ScrollReveal>
-          </Heading>
+          <Heading as="h1" size="lg" className="mb-8 ml-28">FACULTY COORDINATOR</Heading>
           <div className="w-full flex justify-center mb-12">
             <div className="w-full max-w-sm m-auto">
               <CoreTeamInfoCard {...facultyCoordinator} />
@@ -36,11 +31,7 @@ const TeamPage = () => {
           </div>
 
           {/* 👥 Core Team Section */}
-          <Heading as="h1" size="lg" className="mb-8 ml-28">
-            <ScrollReveal>
-              CORE TEAM
-            </ScrollReveal>
-          </Heading>
+          <Heading as="h1" size="lg" className="mb-8 ml-28">CORE TEAM</Heading>
           <div className="w-full flex justify-center">
             <div className="w-[82vw] grid grid-cols-1 md:grid-cols-2 lg:grid-cols-4 gap-8 justify-items-center mb-8">
               {CoreTeam.map((admin, index) => (
@@ -52,11 +43,7 @@ const TeamPage = () => {
           </div>
 
           {/* 🧑‍💻 Team Section */}
-          <Heading as="h1" size="lg" className="mb-8 ml-28">
-            <ScrollReveal>
-              TEAM
-            </ScrollReveal>
-          </Heading>
+          <Heading as="h1" size="lg" className="mb-8 ml-28">TEAM</Heading>
           <div className="w-full flex justify-center">
             <div className="w-[82vw] grid grid-cols-1 md:grid-cols-2 lg:grid-cols-4 gap-8 justify-items-center mb-8">
               {Team.map((admin, index) => (
